src/services/bot/requirements.txt
python-dotenv
redis
aiogram